requirements.txt
fastapi
uvicorn
flake8
google-cloud-firestore
loguru
openai
langchain